UBUNTU: SAUCE: apparmor: Fix new to old label comparison for domain transitions
For the purposes of inherit we should be treating a profile/label transition
to its replacement as if the replacement is the profile/label.

So make the comparison based off of the label proxy, not the label itself.

BugLink: http://bugs.launchpad.net/bugs/1615880

Signed-off-by: John Johansen <john.johansen@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Kamal Mostafa <kamal@canonical.com>
(cherry picked from commit 52626f9c7a68f7a22ca16aab4629e5eab31abba5)
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>